refactor(extension): separate settings ownership
